Relocate copied content
And get rid of the test-case. IMO it's over-complicated and the new docstring tests are preferable.
