+++ pyiron_snippets/colors.py
