+++ pyiron_snippets/dotdict.py
