+++ pyiron_snippets/files.py
