+++ pyiron_snippets/singleton.py
