--- pyiron_workflow/snippets/__init__.py
@@ -1,8 +1,0 @@
-"""
-Bits of useful pyiron code collected from all over (with git history preserved).
-
-Centralized here to reduce dependencies.
-
-Intended for their own repo that is lightweight and only depends on the standard
-library.
-"""
