--- tests/benchmark/test_benchmark.py
@@ -1,6 +1,0 @@
-import unittest
-
-
-class TestNothing(unittest.TestCase):
-    def test_nothing(self):
-        self.assertTrue(True)
