+++ tests/unit/test_factory.py
@@ -7,7 +7,7 @@ import unittest
 
 import cloudpickle
 
-from pyiron_workflow.snippets.factory import (
+from pyiron_snippets.factory import (
     _ClassFactory,
     _FactoryMade,
     ClassFactory,
